+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - #97 pull request is merged, initial implementation of peewee 3.5+ support thanks to @Koos85
 - Dropped support of peewee <= 3.5 and Python <= 3.5
+- Tests are rewritten to async / await syntax
 
 ## 0.5.12